Separate close and abort actions
Split out the close socket action into a separate close method from the
abort method.  The closure of the socket means that the client can't be
connected again as it doesn't have a socket any more to communicate
over.  Abort now means cancel the application level connection, but
leave the socket open for future connections.