Use plain functions for serialization
Refactor streaming interface to use plain reader functions that take a
stream as input and returns the read value and writer functions that
take a stream and a value as input and writes it to the stream.  This
unifies the interface for read/write of all value types enabling them to
be more easily and consistently composed.